Relocate tatr detector config
